Refactoring: Extracted PlayerBar logic into its own class
