+++ Project/core/src/main/java/View/screen/GameScreen.java
@@ -2,24 +2,17 @@ package View.screen;
 
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.graphics.g2d.TextureRegion;
-import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
 
-import Model.Player;
-
-import java.util.ArrayList;
-import java.util.List;
+import View.screen.GameScreenComponents.PlayerBar;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.ui.Value;
-import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.scenes.scene2d.ui.Image;
 import io.github.MindMGMT.MindMGMT;
@@ -35,7 +28,7 @@ public class GameScreen implements Screen {
     private final SpriteBatch batch;
     private final Texture boardTexture;
     private final Image boardImage;
-    private int currentTurn; //TODO: Remove
+    private final PlayerBar playerBar;
 
 
     public GameScreen(MindMGMT game) {
@@ -47,7 +40,7 @@ public class GameScreen implements Screen {
         this.skin = new Skin(Gdx.files.internal("metalui/metal-ui.json"));
         this.boardTexture = new Texture("basic-board.png");
         this.boardImage = new Image(boardTexture);
-        this.currentTurn = 0; //TODO: Remove
+        this.playerBar = new PlayerBar(game.nrOfPayers, skin);
 
         Gdx.input.setInputProcessor(stage);
         setupUI();
@@ -64,47 +57,7 @@ public class GameScreen implements Screen {
     }
 
     private void setupPlayerBar(Table root) {
-        Table playerBar = new Table();
         root.add(playerBar).expandX().fillX().top().height(stage.getViewport().getWorldHeight() * 0.1f);
-
-        // Buttons for each player
-        // TODO: Update to use real model and not fake data
-        List<TextButton> playerButtons = new ArrayList<>();
-        for (int i = 0; i < game.nrOfPayers; i++) {
-            TextButton playerButton = new TextButton("Player " + (i + 1), skin);
-            if (i == currentTurn) {
-                playerButton.setColor(Color.GREEN);
-            }
-            playerButtons.add(playerButton);
-            playerBar.add(playerButton).expandX();
-        }
-
-        // Button for simulating that it's the next players turn
-        // TODO: Remove button below
-        TextButton nextTurnButton = new TextButton("Next turn", skin);
-        nextTurnButton.setColor(Color.MAGENTA);
-        playerBar.add(nextTurnButton);
-        nextTurnButton.addListener(new ChangeListener() {
-            @Override
-            public void changed(ChangeEvent event, Actor actor) {
-                playerButtons.get(currentTurn).setColor(Color.WHITE);
-                currentTurn = (currentTurn + 1) % game.nrOfPayers;
-                playerButtons.get(currentTurn).setColor(Color.GREEN);
-            }
-        });
-
-        //get the player list from the game controller
-        //there is no getgame method in gamecontroller yet, and not sure if it's needed
-        // for (Player player : gameController.getGame().getPlayers()) {
-        //     String playerName = player.getName();
-        //     TextButton playerButton = new TextButton(playerName, skin);
-        //     playerBar.add(playerButton).expandX();
-
-        //     //highlight the current player
-        //     if (player.equals(gameController.getCurrentPlayer())) {
-        //         playerButton.setColor(0, 1, 0, 1);
-        //     }
-        // }
     }
 
     private void setupMainSection(Table root) {
